requirements.txt
discord.py>=2.0.0
groq>=0.4.0
openai>=1.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
